pyproject.toml
[tool.poetry]
name = "obsidian-agent"
packages = [{ include = "obsidian_agent", from = "src" }]
version = "0.1.0"
description = "A tool for analyzing Obsidian vaults and generating graph representations of note connections."
authors = ["Kenneth Ezekiel Suprantoni <your.email@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/KenEzekiel/obsidian-agent"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
rich = "^14.0.0"
langchain = "^0.3.25"
langgraph = "^0.4.8"
chromadb = "^1.0.12"
pydantic = "^2.11.5"
pyyaml = "^6.0.2"
sentence-transformers = "^4.1.0"
openai = "^1.85.0"
tiktoken = "^0.9.0"
langchain-google-genai = "^2.1.5"

[tool.poetry.scripts]
obsidian-agent = "obsidian_agent.scripts.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
